Updated TCMP - Now follows algorithm more closely
TCMP is tested to work with constraint on y-axis, and it should be able to translate to constraint on x or y axis.

It should also be able to handle multiple constraints (xy, xz, and yz) but that is not tested yet.

Added simple constraint to tree that it should be above table-plane (i.e. z>0)

Need to find other ways to optimize, as collision-detect takes most of the time.